Working GUI Build with Document Finder

+++ Sprint5/page.cpp
@@ -8,7 +8,7 @@ Page::Page()
 
 }
 
-Page::Page(float i, std::string t, std::string b, std::string c){
+Page::Page(int i, std::string t, std::string b, std::string c){
     Id = i;
     title = t;
     body = b;
